[Graphs] Add path finding algorithm interface and Dijkstra algorithm first implementation. Still not takes into account available time

+++ SupportSDK/Graphs/Connection.h
@@ -13,16 +13,16 @@ namespace SDK
 			private:
 				GraphNode*	mp_from;
 				GraphNode*	mp_to;
-				int					m_cost;
+				float				m_cost;
 
 			public:
-				Connection(GraphNode* ip_from_node, GraphNode* ip_to_node, int i_cost)
+				Connection(GraphNode* ip_from_node, GraphNode* ip_to_node, float i_cost)
 					: mp_from(ip_from_node)
 					, mp_to(ip_to_node)
 					, m_cost(i_cost)
 					{}
 				virtual ~Connection(){}
-				int					GetCost() const { return m_cost; }
+				float				GetCost() const { return m_cost; }
 				GraphNode*	GetFromNode() const { return mp_from; }
 				GraphNode*	GetToNode() const { return mp_to; }
 			};
